submission/code/requirements.txt
torch==2.1.0
torchvision==0.16.0
opencv-python==4.9.0.80
numpy==1.26.3
Pillow==10.2.0